faltan 3 test

+++ src/Cine.Core/Entrada.cs
@@ -5,9 +5,9 @@ namespace Cine.Core
         public byte NumEntrada { get; set; }
         public ushort IdProyeccion { get; set; }
         public byte IdCliente { get; set; }
-        public float Valor { get; set; }
+        public decimal Valor { get; set; }
 
-        public Entrada(byte NumEntrada, ushort IdProyeccion, byte IdCliente, float valor)
+        public Entrada(byte NumEntrada, ushort IdProyeccion, byte IdCliente, decimal valor)
         {
             this.NumEntrada = NumEntrada;
             this.IdProyeccion = IdProyeccion;
